LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.lineTo.nonfinite-expected.txt

2d.path.lineTo.nonfinite

lineTo() with Infinity/NaN is ignored


PASS lineTo() with Infinity/NaN is ignored 
